requirements.txt
flask>=2.3.0
requests>=2.31.0
chromadb>=0.4.0
PyYAML>=6.0.1
pytest>=8.0.0
